+++ apps/api/src/app/portfolio/calculator/roai/portfolio-calculator-cash.spec.ts
@@ -237,6 +237,7 @@ describe('PortfolioCalculator', () => {
         dateOfFirstActivity: '2023-12-31',
         dividend: new Big(0),
         dividendInBaseCurrency: new Big(0),
+        dividendYieldTrailingTwelveMonths: 0,
         fee: new Big(0),
         feeInBaseCurrency: new Big(0),
         grossPerformance: new Big(0),