+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -1865,6 +1865,7 @@ export class PortfolioService {
 
     const {
       currentValueInBaseCurrency,
+      dividendYieldTrailingTwelveMonths,
       totalInvestment,
       totalInvestmentWithCurrencyEffect
     } = await portfolioCalculator.getSnapshot();
@@ -1969,6 +1970,7 @@ export class PortfolioService {
       annualizedPerformancePercentWithCurrencyEffect,
       cash,
       dateOfFirstActivity,
+      dividendYieldTrailingTwelveMonths,
       excludedAccountsAndActivities,
       netPerformance,
       netPerformancePercentage,